Added a very naive bi-morphic inline cache
The inline cache is naive because I just keep an array for the classes as well as an array for the cached invokables, both with the same length as the bytecode array. These arrays are indexed by the bytecode offset, which allows to safely cache two invokables. The first one at the position of the send bytecode, and the second one at the send-bytecode's parameter (the selector). This encoding is not very efficient, because it wastes quite a bit of memory since the arrays will be rather sparsely populated.

Signed-off-by: Stefan Marr <git@stefan-marr.de>